Users with some draft permissions can access the admin pages
